# TestAddon: Merge viteFinal config into vitest config