@yannbf 2024-12-04
@ndelangen I'm not sure if the plugins from frameworks will be included (like a Vite plugin probably can't add other plugins), but if they are, we need to test if this works for svelte, vue3, react-native-web-vite and nextjs without the need of setting the vite plugin like it is described currently in the docs:
https://storybook.js.org/docs/writing-tests/test-addon#framework-plugins

So if it works then:
- the docs need to be reworked
- the postinstall script needs to not do the logic to inject those framework plugins as it does not
- we need to make sure that we notify users that their configs need to change (the framework plugins would not be needed)