@JReinhold 2024-12-04
This limitation is going to hurt us:

![image](https://github.com/user-attachments/assets/9ea6505f-0d32-4ca8-8434-c312aadf9763)

from https://vite.dev/guide/api-plugin.html#config

We can modify the config as much as we want, but any modifications to the plugins array won't have an affect. Internally we make heavy use of adding Vite-specifics via plugins, and it's also the recommended pattern in userland, so this is not an edge case. It means that even with this work here, we're _still_ not getting the actual Storybook Vite configuration into the Vitest runtime. And so I wonder, if it's best to try or not?

1. Either we keep this: https://storybook.js.org/docs/writing-tests/test-addon#how-do-i-apply-custom-vite-configuration
2. ... or we say, _"**Most** of your `viteFinal` is respected, except for `plugins`, you still have to copy them over to your Vitest config"_