@kylegach 2024-12-04
> * Either we keep this: [storybook.js.org/docs/writing-tests/test-addon#how-do-i-apply-custom-vite-configuration](https://storybook.js.org/docs/writing-tests/test-addon#how-do-i-apply-custom-vite-configuration)
> * ... or we say, _"**Most** of your `viteFinal` is respected, except for `plugins`, you still have to copy them over to your Vitest config"_

From what I've been able to observe, most of the `viteFinal` that is needed is aliases and resolvers. So I think we should keep those docs, but revise them to make it clear that it's only necessary for `plugins`.