@JReinhold 2024-12-06
There might actually be a way around the plugins limitations I explained above.

We can't add plugins in the `config` hook, but we _can_ make the overall `storybookTest` function return multiple plugins. So we could resolve the `viteFinal` preset upfront, find all `plugins` returned, and return them from the function, next to our plugin - and then the rest of the config would be in added in the `config` hook.

We don't need to do this now, let's ship what we have, but this could be a good solution when working on https://github.com/storybookjs/addon-svelte-csf/issues/213